feat(dialog): add reusable confirmation, error, and info dialogs
