+++ src/main/java/org/ndungutse/ems/AppContext.java
@@ -9,20 +9,4 @@ public class AppContext {
     public static EmployeeCollection<Integer> getEmployeeCollection() {
         return employeeCollection;
     }
-
-    public static void showAlert(String title, String content) {
-        Alert alert = new Alert(Alert.AlertType.INFORMATION);
-        alert.setTitle(title);
-        alert.setHeaderText(null);
-        alert.setContentText(content);
-        alert.showAndWait();
-    }
-
-    public static void showErrorAlert(String title, String content) {
-        Alert alert = new Alert(Alert.AlertType.ERROR);
-        alert.setTitle(title);
-        alert.setHeaderText(null);
-        alert.setContentText(content);
-        alert.showAndWait();
-    }
 }
