+++ src/main/java/org/ndungutse/ems/HelloApplication.java
@@ -16,6 +16,8 @@ import org.ndungutse.ems.repository.EmployeeCollection;
 public class HelloApplication extends Application {
         @Override
         public void start(Stage stage) throws IOException {
+
+                try {
                 // Load the FXML file
                 FXMLLoader fxmlLoader = new FXMLLoader(HelloApplication.class.getResource("hello-view.fxml"));
 
@@ -30,6 +32,10 @@ public class HelloApplication extends Application {
 
                 // Show the stage
                 stage.show();
+
+                }catch (IOException e ){
+                        System.out.println(e.getMessage());
+                }
         }
         public static final EmployeeCollection<Integer> employeeCollection = AppContext.getEmployeeCollection();
